requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
PyYAML==6.0.1
